+++ src/uni_modules/uview-plus/components/u-swipe-action-item/props.js
@@ -29,6 +29,11 @@ export const props = defineMixin({
             type: Boolean,
             default: () => defProps.swipeActionItem.autoClose
         },
+        // 是否正在横向滑动，用于外部通过v-model:scrolling暂停页面或容器滚动
+        scrolling: {
+            type: Boolean,
+            default: () => defProps.swipeActionItem.scrolling
+        },
         // 滑动距离阈值，只有大于此值，才被认为是要打开菜单
         threshold: {
             type: Number,